pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "track-tram-reliability"
version = "0.1.0"
description = "Collect and store reliability data (delays, cancellations) for MVG public transportation"
authors = [
  { name = "Rovo Dev" }
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "requests>=2.32.0",
  "pydantic>=2.7.0",
  "SQLAlchemy>=2.0.0",
  "alembic>=1.13.0",
  "PyYAML>=6.0.0",
  "typer[all]>=0.12.0",
]

[project.urls]
Homepage = "https://example.com/track-tram-reliability"

[project.scripts]
"ttr" = "track_tram_reliability.cli:app"

[tool.setuptools]
packages = ["track_tram_reliability"]
package-dir = {"track_tram_reliability" = "src/track_tram_reliability"}

[tool.ruff]
line-length = 100
